Add `ffretag` helper script
This can be used to set a piece of metadata replacing the file in-line.

By default `ffmpeg` won't let you set the output file the same as the
input file; this is a simple script built on top of `ffmpeg` that allows
this using `/tmp/` to hold the file temporarily.